refactor(renovate-changesets): remove superseded detection modules
