--- .github/actions/renovate-changesets/src/detectors/security-patterns.ts
@@ -1,105 +1,0 @@
-import type {SecurityPatterns} from './security-vulnerability-types.js'
-
-export const SECURITY_PATTERNS: SecurityPatterns = {
-  npm: {
-    vulnerabilityKeywords: [
-      'remote code execution',
-      'rce',
-      'arbitrary code execution',
-      'prototype pollution',
-      'cross-site scripting',
-      'xss',
-      'sql injection',
-      'path traversal',
-      'directory traversal',
-      'deserialization',
-      'command injection',
-      'server-side request forgery',
-      'ssrf',
-      'denial of service',
-      'dos',
-      'privilege escalation',
-      'information disclosure',
-      'crypto',
-      'cryptographic',
-      'buffer overflow',
-      'memory corruption',
-    ],
-    exploitKeywords: [
-      'exploit',
-      'poc',
-      'proof of concept',
-      'active exploitation',
-      'in the wild',
-      'zero day',
-      '0day',
-      'malicious',
-      'backdoor',
-      'trojan',
-      'supply chain attack',
-    ],
-    riskPackages: [
-      'eval',
-      'vm2',
-      'serialize-javascript',
-      'node-serialize',
-      'safer-eval',
-      'express-fileupload',
-      'multer',
-    ],
-    trustedSources: [
-      '@types/',
-      '@angular/',
-      '@babel/',
-      '@typescript-eslint/',
-      '@react/',
-      '@vue/',
-      'react',
-      'vue',
-      'express',
-      'lodash',
-    ],
-  },
-  docker: {
-    vulnerabilityKeywords: [
-      'container escape',
-      'privilege escalation',
-      'root access',
-      'host access',
-      'kernel vulnerability',
-      'runtime vulnerability',
-      'image vulnerability',
-      'base image',
-      'malicious image',
-    ],
-    exploitKeywords: [
-      'container breakout',
-      'docker escape',
-      'runtime exploit',
-      'privilege escalation',
-      'root compromise',
-    ],
-    riskPackages: ['alpine', 'ubuntu', 'debian', 'centos', 'node', 'python'],
-    trustedSources: ['official', 'library/', 'alpine', 'ubuntu', 'debian'],
-  },
-  'github-actions': {
-    vulnerabilityKeywords: [
-      'token injection',
-      'script injection',
-      'workflow injection',
-      'secrets exposure',
-      'privilege escalation',
-      'runner compromise',
-      'malicious action',
-    ],
-    exploitKeywords: [
-      'repository takeover',
-      'secrets theft',
-      'workflow manipulation',
-      'runner exploitation',
-      'supply chain attack',
-    ],
-    riskPackages: ['actions/', 'third-party actions', 'community actions'],
-    trustedSources: ['actions/', 'github/', 'microsoft/', 'azure/'],
-  },
-}
