--- .github/actions/renovate-changesets/src/security-vulnerability-detector.ts
@@ -1,76 +1,0 @@
-import type {Octokit} from '@octokit/rest'
-import type {
-  SecurityAnalysis,
-  SecurityVulnerability,
-} from './detectors/security-vulnerability-types.js'
-import type {RenovateDependency} from './renovate-parser.js'
-import {
-  extractCVEVulnerabilities,
-  extractGHSAVulnerabilities,
-} from './detectors/security-advisory-parser.js'
-import {
-  analyzePRContentSecurity,
-  analyzeSecurityPatterns,
-  analyzeSupplyChainRisks,
-} from './detectors/security-change-analyzer.js'
-import {SECURITY_PATTERNS} from './detectors/security-patterns.js'
-import {synthesizeSecurityAnalysis} from './detectors/security-severity-classifier.js'
-
-export type {
-  CVEInfo,
-  GHSAInfo,
-  SecurityAnalysis,
-  SecurityPatterns,
-  SecurityVulnerability,
-} from './detectors/security-vulnerability-types.js'
-
-export interface SecurityVulnerabilityAnalysisOptions {
-  prContentOrOctokit?: string | Octokit
-  owner?: string
-  repo?: string
-  prNumber?: number
-  securityPatterns?: typeof SECURITY_PATTERNS
-}
-
-export async function analyzeSecurityVulnerabilities(
-  dependency: RenovateDependency,
-  options: SecurityVulnerabilityAnalysisOptions = {},
-): Promise<SecurityAnalysis> {
-  const {prContentOrOctokit, owner, repo, prNumber, securityPatterns = SECURITY_PATTERNS} = options
-
-  const vulnerabilities: SecurityVulnerability[] = dependency.isSecurityUpdate
-    ? [
-        {
-          id: 'RENOVATE_SECURITY_FLAG',
-          type: 'security_flag',
-          severity: 'medium',
-          description: 'Renovate security update flag indicates potential security fix',
-          source: 'renovate',
-          affectedVersions: dependency.currentVersion == null ? [] : [dependency.currentVersion],
-        },
-      ]
-    : []
-
-  vulnerabilities.push(...extractCVEVulnerabilities(dependency))
-  vulnerabilities.push(...extractGHSAVulnerabilities(dependency))
-  vulnerabilities.push(...analyzeSecurityPatterns(dependency, securityPatterns))
-  vulnerabilities.push(...analyzeSupplyChainRisks(dependency, securityPatterns))
-
-  if (
-    prContentOrOctokit != null &&
-    typeof prContentOrOctokit !== 'string' &&
-    owner != null &&
-    repo != null &&
-    prNumber != null
-  ) {
-    try {
-      vulnerabilities.push(
-        ...(await analyzePRContentSecurity(dependency, prContentOrOctokit, owner, repo, prNumber)),
-      )
-    } catch (error) {
-      console.warn('Failed to analyze PR content for security:', error)
-    }
-  }
-
-  return synthesizeSecurityAnalysis(vulnerabilities)
-}
