--- .github/actions/renovate-changesets/src/summaries/js-ecosystem-summaries.ts
@@ -1,123 +1,0 @@
-import type {RenovatePRContext} from '../renovate-parser'
-import type {ImpactAssessment} from '../semver-impact-assessor'
-import type {SummaryGeneratorConfig} from '../summary-generator-types'
-
-import {formatVersionText} from './summary-helpers'
-
-export interface JsEcosystemSummaryContext {
-  config: Pick<
-    SummaryGeneratorConfig,
-    | 'includeBreakingChangeWarnings'
-    | 'includeVersionDetails'
-    | 'maxDependenciesToList'
-    | 'sortDependencies'
-  >
-  addBreakingChangeWarning: (summary: string, impactAssessment: ImpactAssessment) => string
-  generateGroupedUpdateSummary: (
-    managerType: string,
-    dependencies: string[],
-    impactAssessment: ImpactAssessment,
-  ) => string
-  generateSecurityUpdateSummary: (
-    managerType: string,
-    dependencies: string[],
-    prContext: RenovatePRContext,
-    impactAssessment: ImpactAssessment,
-  ) => string
-  generateSingleDependencySummary: (
-    managerType: string,
-    dependency: string,
-    prContext: RenovatePRContext,
-    impactAssessment: ImpactAssessment,
-    emoji: string,
-  ) => string
-  getEmojiForUpdate: (prContext: RenovatePRContext, impactAssessment: ImpactAssessment) => string
-}
-
-export function generateNpmSummaryLogic(
-  ctx: JsEcosystemSummaryContext,
-  prContext: RenovatePRContext,
-  impactAssessment: ImpactAssessment,
-  dependencies: string[],
-): string {
-  const emoji = ctx.getEmojiForUpdate(prContext, impactAssessment)
-  const sortedDeps = ctx.config.sortDependencies ? [...dependencies].sort() : dependencies
-
-  if (prContext.isSecurityUpdate) {
-    return ctx.generateSecurityUpdateSummary('npm', sortedDeps, prContext, impactAssessment)
-  }
-
-  if (prContext.isGroupedUpdate) {
-    return ctx.generateGroupedUpdateSummary('npm', sortedDeps, impactAssessment)
-  }
-
-  if (sortedDeps.length === 0) {
-    return `${emoji}Update npm dependencies`
-  }
-
-  if (sortedDeps.length === 1) {
-    const dependency = sortedDeps[0]
-    if (dependency) {
-      return ctx.generateSingleDependencySummary(
-        'npm',
-        dependency,
-        prContext,
-        impactAssessment,
-        emoji,
-      )
-    }
-  }
-
-  if (sortedDeps.length <= ctx.config.maxDependenciesToList) {
-    const depList = sortedDeps.map(dep => `\`${dep}\``).join(', ')
-    const summary = `${emoji}Update npm dependencies: ${depList}`
-
-    if (impactAssessment.hasBreakingChanges && ctx.config.includeBreakingChangeWarnings) {
-      return `${summary}\n\n⚠️ **Breaking Changes**: This update includes breaking changes that may require code modifications.`
-    }
-
-    return summary
-  }
-
-  const summary = `${emoji}Update ${sortedDeps.length} npm dependencies`
-  return ctx.addBreakingChangeWarning(summary, impactAssessment)
-}
-
-export function generateGitHubActionsSummaryLogic(
-  ctx: JsEcosystemSummaryContext,
-  prContext: RenovatePRContext,
-  impactAssessment: ImpactAssessment,
-  dependencies: string[],
-): string {
-  const emoji = ctx.getEmojiForUpdate(prContext, impactAssessment)
-  const sortedDeps = ctx.config.sortDependencies ? [...dependencies].sort() : dependencies
-
-  if (prContext.isSecurityUpdate) {
-    return ctx.generateSecurityUpdateSummary(
-      'GitHub Actions',
-      sortedDeps,
-      prContext,
-      impactAssessment,
-    )
-  }
-
-  if (sortedDeps.length === 1) {
-    const dep = sortedDeps[0]
-    const versionInfo = prContext.dependencies.find(d => d.name === dep)
-    const versionText = formatVersionText(
-      versionInfo?.currentVersion,
-      versionInfo?.newVersion,
-      impactAssessment.overallImpact,
-      ctx.config.includeVersionDetails,
-    )
-
-    return `${emoji}Update GitHub Actions workflow dependency \`${dep}\`${versionText}`
-  }
-
-  if (sortedDeps.length <= ctx.config.maxDependenciesToList) {
-    const depList = sortedDeps.map(dep => `\`${dep}\``).join(', ')
-    return `${emoji}Update GitHub Actions workflow dependencies: ${depList}`
-  }
-
-  return `${emoji}Update ${sortedDeps.length} GitHub Actions workflow dependencies`
-}
