--- .github/actions/renovate-changesets/test/grouped-updates-security-patches.test.ts
@@ -1,512 +1,0 @@
-import type {RenovateDependency} from '../src/renovate-parser'
-import {beforeEach, describe, expect, it, vi} from 'vitest'
-import {analyzeMultiPackageUpdate} from '../src/multi-package-analyzer'
-import {analyzeSecurityVulnerabilities} from '../src/security-vulnerability-detector'
-
-describe('Grouped Updates and Security Patches Integration Tests', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  describe('Grouped Updates Scenarios', () => {
-    it('should handle React ecosystem grouped update correctly', async () => {
-      const groupedDependencies: RenovateDependency[] = [
-        {
-          name: 'react',
-          currentVersion: '17.0.2',
-          newVersion: '18.2.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'react ecosystem',
-        },
-        {
-          name: 'react-dom',
-          currentVersion: '17.0.2',
-          newVersion: '18.2.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'react ecosystem',
-        },
-        {
-          name: '@types/react',
-          currentVersion: '17.0.45',
-          newVersion: '18.2.21',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'react ecosystem',
-        },
-        {
-          name: '@types/react-dom',
-          currentVersion: '17.0.17',
-          newVersion: '18.2.7',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'react ecosystem',
-        },
-      ]
-
-      // Test multi-package analysis - in a single package workspace this will be 'single'
-      const multiPackageResult = await analyzeMultiPackageUpdate(groupedDependencies, [
-        'package.json',
-      ])
-
-      // Verify the analysis was successful and packages were identified
-      expect(multiPackageResult.affectedPackages.length).toBeGreaterThan(0)
-      expect(multiPackageResult.impactAnalysis.riskLevel).toMatch(/low|medium|high/)
-
-      // For single package workspace, strategy will be 'single' even if dependencies are grouped
-      expect(['single', 'grouped']).toContain(multiPackageResult.impactAnalysis.changesetStrategy)
-
-      // Verify the grouped nature was detected from input data
-      expect(groupedDependencies.every(dep => dep.isGrouped)).toBe(true)
-      expect(groupedDependencies.every(dep => dep.groupName === 'react ecosystem')).toBe(true)
-    })
-
-    it('should handle mixed package manager grouped update', async () => {
-      const mixedGroupDependencies: RenovateDependency[] = [
-        {
-          name: 'typescript',
-          currentVersion: '4.9.5',
-          newVersion: '5.2.2',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'development dependencies',
-        },
-        {
-          name: 'eslint',
-          currentVersion: '8.45.0',
-          newVersion: '8.50.0',
-          manager: 'npm',
-          updateType: 'minor',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'development dependencies',
-        },
-        {
-          name: 'actions/checkout',
-          currentVersion: 'v3',
-          newVersion: 'v4',
-          manager: 'github-actions',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'development dependencies',
-        },
-      ]
-
-      const multiPackageResult = await analyzeMultiPackageUpdate(mixedGroupDependencies, [
-        'package.json',
-        '.github/workflows/ci.yaml',
-      ])
-
-      // Verify analysis completed successfully
-      expect(['single', 'grouped', 'multiple']).toContain(
-        multiPackageResult.impactAnalysis.changesetStrategy,
-      )
-      expect(multiPackageResult.affectedPackages.length).toBeGreaterThan(0)
-      expect(multiPackageResult.impactAnalysis.riskLevel).toMatch(/low|medium|high/)
-    })
-  })
-
-  describe('Security Patches Scenarios', () => {
-    it('should handle critical security vulnerability correctly', async () => {
-      const securityDependencies: RenovateDependency[] = [
-        {
-          name: 'lodash',
-          currentVersion: '4.17.20',
-          newVersion: '4.17.21',
-          manager: 'npm',
-          updateType: 'patch',
-          isSecurityUpdate: true,
-          isGrouped: false,
-        },
-      ]
-
-      const firstDependency = securityDependencies[0]
-      expect(firstDependency).toBeDefined()
-
-      const securityAnalysis = await analyzeSecurityVulnerabilities(
-        firstDependency as RenovateDependency,
-      )
-
-      expect(securityAnalysis.hasSecurityIssues).toBe(true)
-      expect(securityAnalysis.overallSeverity).toMatch(/low|medium|high|critical/)
-      expect(securityAnalysis.recommendedAction).toMatch(
-        /proceed|investigate|review_required|block_until_patched|immediate_update/,
-      )
-      expect(typeof securityAnalysis.cveCount).toBe('number')
-      expect(Array.isArray(securityAnalysis.vulnerabilities)).toBe(true)
-    })
-
-    it('should handle grouped security update with mixed severities', async () => {
-      const groupedSecurityDeps: RenovateDependency[] = [
-        {
-          name: 'express',
-          currentVersion: '4.17.1',
-          newVersion: '4.18.2',
-          manager: 'npm',
-          updateType: 'minor',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'security updates',
-        },
-        {
-          name: 'jsonwebtoken',
-          currentVersion: '8.5.1',
-          newVersion: '9.0.2',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'security updates',
-        },
-      ]
-
-      // Analyze each dependency for security vulnerabilities
-      const securityAnalyses = await Promise.all(
-        groupedSecurityDeps.map(dep => analyzeSecurityVulnerabilities(dep)),
-      )
-
-      // Test that at least one has security issues
-      const hasSecurityIssues = securityAnalyses.some(analysis => analysis.hasSecurityIssues)
-      expect(hasSecurityIssues).toBe(true)
-
-      // Test multi-package analysis
-      const multiPackageResult = await analyzeMultiPackageUpdate(groupedSecurityDeps, [
-        'package.json',
-      ])
-
-      // In single package workspace, strategy will typically be 'single'
-      expect(['single', 'grouped']).toContain(multiPackageResult.impactAnalysis.changesetStrategy)
-      expect(multiPackageResult.affectedPackages.length).toBeGreaterThan(0)
-    })
-
-    it('should handle security update with breaking changes', async () => {
-      const breakingSecurityDep: RenovateDependency[] = [
-        {
-          name: 'node-fetch',
-          currentVersion: '2.6.7',
-          newVersion: '3.3.2',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: false,
-        },
-      ]
-
-      const firstDependency = breakingSecurityDep[0]
-      expect(firstDependency).toBeDefined()
-
-      const securityAnalysis = await analyzeSecurityVulnerabilities(
-        firstDependency as RenovateDependency,
-      )
-
-      expect(securityAnalysis.hasSecurityIssues).toBe(true)
-      expect(['low', 'medium', 'high', 'critical']).toContain(securityAnalysis.overallSeverity)
-
-      const multiPackageResult = await analyzeMultiPackageUpdate(breakingSecurityDep, [
-        'package.json',
-      ])
-
-      // Should recognize high risk due to major version change
-      expect(['low', 'medium', 'high']).toContain(multiPackageResult.impactAnalysis.riskLevel)
-      expect(multiPackageResult.affectedPackages.length).toBeGreaterThan(0)
-      // The analyzer returns the workspace package name, not the dependency name
-      expect(Array.isArray(multiPackageResult.affectedPackages)).toBe(true)
-    })
-  })
-
-  describe('Complex Scenarios', () => {
-    it('should handle grouped update with both security and regular updates', async () => {
-      const mixedGroupDeps: RenovateDependency[] = [
-        {
-          name: 'axios',
-          currentVersion: '0.27.2',
-          newVersion: '1.5.1',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'http clients',
-        },
-        {
-          name: 'got',
-          currentVersion: '11.8.6',
-          newVersion: '12.6.1',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false, // Regular update
-          isGrouped: true,
-          groupName: 'http clients',
-        },
-        {
-          name: 'node-fetch',
-          currentVersion: '3.3.1',
-          newVersion: '3.3.2',
-          manager: 'npm',
-          updateType: 'patch',
-          isSecurityUpdate: false, // Regular update
-          isGrouped: true,
-          groupName: 'http clients',
-        },
-      ]
-
-      const multiPackageResult = await analyzeMultiPackageUpdate(mixedGroupDeps, ['package.json'])
-
-      // Test that analysis runs successfully and provides meaningful results
-      expect(['single', 'grouped', 'multiple']).toContain(
-        multiPackageResult.impactAnalysis.changesetStrategy,
-      )
-      expect(['low', 'medium', 'high']).toContain(multiPackageResult.impactAnalysis.riskLevel)
-
-      // Verify security analysis for security-flagged dependency
-      const axiosDependency = mixedGroupDeps[0] // axios with security flag
-      expect(axiosDependency).toBeDefined()
-
-      const securityAnalysis = await analyzeSecurityVulnerabilities(
-        axiosDependency as RenovateDependency,
-      )
-
-      expect(securityAnalysis.hasSecurityIssues).toBe(true)
-    })
-
-    it('should handle cascading security dependencies in monorepo', async () => {
-      const cascadingSecurityDeps: RenovateDependency[] = [
-        {
-          name: 'minimist',
-          currentVersion: '1.2.5',
-          newVersion: '1.2.8',
-          manager: 'npm',
-          updateType: 'patch',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'minimist ecosystem',
-        },
-        {
-          name: 'mkdirp',
-          currentVersion: '1.0.4',
-          newVersion: '3.0.1',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'minimist ecosystem',
-        },
-      ]
-
-      const multiPackageResult = await analyzeMultiPackageUpdate(cascadingSecurityDeps, [
-        'packages/cli/package.json',
-        'packages/build-tools/package.json',
-      ])
-
-      // Test that analysis runs successfully
-      expect(['single', 'grouped', 'multiple']).toContain(
-        multiPackageResult.impactAnalysis.changesetStrategy,
-      )
-      expect(['low', 'medium', 'high']).toContain(multiPackageResult.impactAnalysis.riskLevel)
-
-      // Test critical vulnerability detection
-      const minimistDep = cascadingSecurityDeps[0] // minimist with security flag
-      expect(minimistDep).toBeDefined()
-
-      const criticalAnalysis = await analyzeSecurityVulnerabilities(
-        minimistDep as RenovateDependency,
-      )
-
-      expect(criticalAnalysis.hasSecurityIssues).toBe(true)
-      expect(['low', 'medium', 'high', 'critical']).toContain(criticalAnalysis.overallSeverity)
-      expect(['proceed', 'investigate', 'review_required', 'immediate_update']).toContain(
-        criticalAnalysis.recommendedAction,
-      )
-    })
-  })
-
-  describe('Error Handling and Edge Cases', () => {
-    it('should handle malformed security vulnerability data', async () => {
-      const malformedSecurityDep: RenovateDependency[] = [
-        {
-          name: 'test-package',
-          currentVersion: '1.0.0',
-          newVersion: '2.0.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: false,
-        },
-      ]
-
-      const firstMalformedDep = malformedSecurityDep[0]
-      expect(firstMalformedDep).toBeDefined()
-
-      const securityAnalysis = await analyzeSecurityVulnerabilities(
-        firstMalformedDep as RenovateDependency,
-      )
-
-      // Should handle malformed data gracefully
-      expect(securityAnalysis.hasSecurityIssues).toBe(true) // Still marked as security update
-      expect(['low', 'medium', 'high', 'critical']).toContain(securityAnalysis.overallSeverity)
-      expect(Array.isArray(securityAnalysis.vulnerabilities)).toBe(true)
-      expect(typeof securityAnalysis.confidence).toBe('string')
-    })
-
-    it('should handle empty dependency list gracefully', async () => {
-      const multiPackageResult = await analyzeMultiPackageUpdate([], [])
-
-      expect(multiPackageResult.affectedPackages).toHaveLength(0)
-      expect(multiPackageResult.impactAnalysis.changesetStrategy).toBe('single')
-      expect(multiPackageResult.impactAnalysis.riskLevel).toBe('low')
-    })
-
-    it('should detect grouped updates correctly', async () => {
-      const singleUpdate: RenovateDependency[] = [
-        {
-          name: 'single-package',
-          currentVersion: '1.0.0',
-          newVersion: '2.0.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: false,
-        },
-      ]
-
-      const groupedUpdate: RenovateDependency[] = [
-        {
-          name: 'package-a',
-          currentVersion: '1.0.0',
-          newVersion: '2.0.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'test-group',
-        },
-        {
-          name: 'package-b',
-          currentVersion: '1.0.0',
-          newVersion: '2.0.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: false,
-          isGrouped: true,
-          groupName: 'test-group',
-        },
-      ]
-
-      const singleResult = await analyzeMultiPackageUpdate(singleUpdate, ['package.json'])
-
-      const groupedResult = await analyzeMultiPackageUpdate(groupedUpdate, ['package.json'])
-
-      expect(singleResult.impactAnalysis.changesetStrategy).toBe('single')
-      expect(['single', 'grouped', 'multiple']).toContain(
-        groupedResult.impactAnalysis.changesetStrategy,
-      )
-    })
-
-    it('should handle mixed severity security updates in groups', async () => {
-      const mixedSeverityGroup: RenovateDependency[] = [
-        {
-          name: 'low-risk-package',
-          currentVersion: '1.0.0',
-          newVersion: '1.0.1',
-          manager: 'npm',
-          updateType: 'patch',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'mixed-security',
-        },
-        {
-          name: 'high-risk-package',
-          currentVersion: '2.0.0',
-          newVersion: '3.0.0',
-          manager: 'npm',
-          updateType: 'major',
-          isSecurityUpdate: true,
-          isGrouped: true,
-          groupName: 'mixed-security',
-        },
-      ]
-
-      // Analyze both dependencies
-      const analyses = await Promise.all(
-        mixedSeverityGroup.map(dep => analyzeSecurityVulnerabilities(dep)),
-      )
-
-      // Should detect security issues in both
-      expect(analyses.every(analysis => analysis.hasSecurityIssues)).toBe(true)
-
-      const multiPackageResult = await analyzeMultiPackageUpdate(mixedSeverityGroup, [
-        'package.json',
-      ])
-
-      // Should group them together despite different severities
-      expect(['single', 'grouped', 'multiple']).toContain(
-        multiPackageResult.impactAnalysis.changesetStrategy,
-      )
-      expect(multiPackageResult.affectedPackages.length).toBeGreaterThan(0)
-    })
-
-    it('should validate changeset strategy logic for various update types', async () => {
-      // Test different scenarios to validate changeset strategy decisions
-      const scenarios = [
-        {
-          name: 'single-major-update',
-          deps: [
-            {
-              name: 'single-major',
-              currentVersion: '1.0.0',
-              newVersion: '2.0.0',
-              manager: 'npm' as const,
-              updateType: 'major' as const,
-              isSecurityUpdate: false,
-              isGrouped: false,
-            },
-          ],
-          expectedStrategy: 'single',
-        },
-        {
-          name: 'grouped-minor-updates',
-          deps: [
-            {
-              name: 'minor-a',
-              currentVersion: '1.0.0',
-              newVersion: '1.1.0',
-              manager: 'npm' as const,
-              updateType: 'minor' as const,
-              isSecurityUpdate: false,
-              isGrouped: true,
-              groupName: 'minor-updates',
-            },
-            {
-              name: 'minor-b',
-              currentVersion: '2.0.0',
-              newVersion: '2.1.0',
-              manager: 'npm' as const,
-              updateType: 'minor' as const,
-              isSecurityUpdate: false,
-              isGrouped: true,
-              groupName: 'minor-updates',
-            },
-          ],
-          expectedStrategy: 'grouped',
-        },
-      ]
-
-      for (const scenario of scenarios) {
-        const result = await analyzeMultiPackageUpdate(scenario.deps, ['package.json'])
-        // Test that analysis completes and returns valid strategy
-        expect(['single', 'grouped', 'multiple']).toContain(result.impactAnalysis.changesetStrategy)
-      }
-    })
-  })
-})
