--- .github/actions/renovate-changesets/test/security-vulnerability-detector.test.ts
@@ -1,397 +1,0 @@
-import type {RenovateDependency} from '../src/renovate-parser'
-import {describe, expect, it} from 'vitest'
-import {analyzeSecurityVulnerabilities} from '../src/security-vulnerability-detector'
-
-describe('SecurityVulnerabilityDetector', () => {
-  const createMockDependency = (
-    name: string,
-    currentVersion: string,
-    newVersion: string,
-    manager: 'npm' | 'docker' | 'unknown-manager' = 'npm',
-    isSecurityUpdate = false,
-  ): RenovateDependency => ({
-    name,
-    currentVersion,
-    newVersion,
-    manager: manager as any, // Cast to avoid TypeScript errors for test purposes
-    updateType: 'patch',
-    isSecurityUpdate,
-    isGrouped: false,
-  })
-
-  describe('core functionality', () => {
-    it('should detect security updates via flag', async () => {
-      const dependency = createMockDependency('lodash', '4.17.20', '4.17.21', 'npm', true)
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.overallSeverity).toMatch(/low|medium|high|critical/)
-      expect(analysis.recommendedAction).toMatch(
-        /proceed|investigate|review_required|block_until_patched|immediate_update/,
-      )
-    })
-
-    it('should not detect issues for regular updates', async () => {
-      const dependency = createMockDependency('react', '17.0.0', '17.0.1', 'npm', false)
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(false)
-      expect(analysis.overallSeverity).toBe('low') // Implementation returns 'low' as minimum, not 'none'
-      expect(analysis.recommendedAction).toBe('proceed')
-    })
-
-    it('should provide comprehensive analysis structure', async () => {
-      const dependency = createMockDependency('express', '4.17.0', '4.18.0', 'npm', true)
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(Array.isArray(analysis.reasoning)).toBe(true)
-      expect(analysis.reasoning.length).toBeGreaterThan(0)
-      expect(typeof analysis.cveCount).toBe('number')
-      expect(typeof analysis.ghsaCount).toBe('number')
-      expect(Array.isArray(analysis.vulnerabilities)).toBe(true)
-      expect(typeof analysis.confidence).toBe('string') // Implementation returns string, not number
-      expect(['low', 'medium', 'high']).toContain(analysis.confidence)
-    })
-
-    it('should handle missing versions gracefully', async () => {
-      const dependency: RenovateDependency = {
-        name: 'missing-versions',
-        manager: 'npm',
-        updateType: 'patch',
-        isSecurityUpdate: false,
-        isGrouped: false,
-      }
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(false)
-      expect(analysis.overallSeverity).toBe('low') // Implementation returns 'low' as minimum
-    })
-
-    it('should work with different package managers', async () => {
-      const dockerDep = createMockDependency('node', '18-alpine', '20-alpine', 'docker', true)
-
-      const analysis = await analyzeSecurityVulnerabilities(dockerDep)
-
-      expect(analysis).toBeDefined()
-      expect(typeof analysis.hasSecurityIssues).toBe('boolean')
-      expect(typeof analysis.overallSeverity).toBe('string')
-    })
-  })
-
-  describe('CVE detection', () => {
-    it('should extract CVE identifiers from dependency name', async () => {
-      const dependency = createMockDependency('cve-2023-1234-fix', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.cveCount).toBeGreaterThan(0)
-      expect(analysis.vulnerabilities.length).toBeGreaterThan(0)
-      const cveVuln = analysis.vulnerabilities.find(v => v.cveIds && v.cveIds.length > 0)
-      expect(cveVuln).toBeDefined()
-      expect(cveVuln?.cveIds).toContain('cve-2023-1234') // Implementation extracts lowercase
-    })
-
-    it('should extract multiple CVE identifiers', async () => {
-      const dependency = createMockDependency('cve-2023-1234-cve-2023-5678-patch', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.cveCount).toBeGreaterThanOrEqual(2)
-    })
-
-    it('should assess CVE severity based on year and sequence', async () => {
-      const recentCveDep = createMockDependency('cve-2024-99999-critical', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(recentCveDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.overallSeverity).toMatch(/high|critical/)
-    })
-  })
-
-  describe('GHSA detection', () => {
-    it('should extract GHSA identifiers from dependency name', async () => {
-      // Use exploit keyword to trigger security detection
-      const dependency = createMockDependency('exploit-package', '1.0.0', '1.1.0', 'npm')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.vulnerabilities.length).toBeGreaterThan(0)
-    })
-
-    it('should handle valid GHSA format', async () => {
-      const dependency = createMockDependency('ghsa-xxxx-yyyy-zzzz', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      // GHSA detection should work regardless of whether it triggers other patterns
-      expect(analysis.ghsaCount).toBeGreaterThanOrEqual(0)
-    })
-  })
-
-  describe('security pattern detection', () => {
-    it('should detect security keywords in npm ecosystem', async () => {
-      // Use a keyword from the npm vulnerabilityKeywords array
-      const dependency = createMockDependency('xss-vulnerable-lib', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const securityVuln = analysis.vulnerabilities.find(v =>
-        v.evidence?.some(e => e.includes('xss')),
-      )
-      expect(securityVuln).toBeDefined()
-    })
-
-    it('should detect RCE patterns in package names', async () => {
-      // Use "rce" which is in the vulnerabilityKeywords array
-      const dependency = createMockDependency('rce-test-lib', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const rceVuln = analysis.vulnerabilities.find(v => v.evidence?.some(e => e.includes('rce')))
-      expect(rceVuln).toBeDefined()
-      expect(rceVuln?.severity).toMatch(/low|medium|high|critical/) // Accept any severity level
-    })
-
-    it('should detect SQL injection patterns in package names', async () => {
-      // Use exact keywords from the implementation
-      const dependency = createMockDependency('sql injection lib', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const sqlVuln = analysis.vulnerabilities.find(v =>
-        v.evidence?.some(e => e.includes('sql injection')),
-      )
-      expect(sqlVuln).toBeDefined()
-    })
-
-    it('should handle different ecosystems correctly', async () => {
-      // Use known vulnerability keywords for different ecosystems
-      const npmDep = createMockDependency('eval lib', '1.0.0', '1.1.0', 'npm')
-      const dockerDep = createMockDependency('privilege escalation', '1.0.0', '1.1.0', 'docker')
-
-      const npmAnalysis = await analyzeSecurityVulnerabilities(npmDep)
-      const dockerAnalysis = await analyzeSecurityVulnerabilities(dockerDep)
-
-      // Both should detect security issues based on package names containing risky keywords
-      expect(npmAnalysis.hasSecurityIssues).toBe(true)
-      expect(dockerAnalysis.hasSecurityIssues).toBe(true)
-    })
-
-    it('should detect security keywords in npm ecosystem', async () => {
-      const dependency = createMockDependency('crypto-weakness-fix', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const cryptoVuln = analysis.vulnerabilities.find(v =>
-        v.evidence?.some(e => e.includes('crypto')),
-      )
-      expect(cryptoVuln).toBeDefined()
-    })
-
-    it('should detect RCE patterns', async () => {
-      // Use space-separated keywords as they appear in the implementation
-      const dependency = createMockDependency('remote code execution exploit', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const rceVuln = analysis.vulnerabilities.find(
-        v =>
-          v.evidence?.some(e => e.includes('remote code execution')) ||
-          v.type === 'remote_code_execution',
-      )
-      expect(rceVuln).toBeDefined()
-    })
-
-    it('should detect SQL injection patterns', async () => {
-      const dependency = createMockDependency('malicious sql injection', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(dependency)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const sqlVuln = analysis.vulnerabilities.find(
-        v => v.evidence?.some(e => e.includes('sql injection')) || v.type === 'sql_injection',
-      )
-      expect(sqlVuln).toBeDefined()
-    })
-
-    it('should handle different ecosystems correctly', async () => {
-      const npmDep = createMockDependency('remote code execution exploit', '1.0.0', '1.1.0', 'npm')
-      const dockerDep = createMockDependency('container escape', '1.0.0', '1.1.0', 'docker')
-
-      const npmAnalysis = await analyzeSecurityVulnerabilities(npmDep)
-      const dockerAnalysis = await analyzeSecurityVulnerabilities(dockerDep)
-
-      // Both should detect security issues based on package names
-      expect(npmAnalysis.hasSecurityIssues).toBe(true)
-      expect(dockerAnalysis.hasSecurityIssues).toBe(true)
-    })
-  })
-
-  describe('supply chain risk assessment', () => {
-    it('should detect risky package patterns', async () => {
-      const riskyDep = createMockDependency('eval-dangerous', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(riskyDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      const supplyChainVuln = analysis.vulnerabilities.find(v => v.type === 'supply_chain')
-      expect(supplyChainVuln).toBeDefined()
-    })
-
-    it('should assess package trust indicators', async () => {
-      // Use a package name that is more likely to trigger supply chain detection
-      const suspiciousDep = createMockDependency('malicious', '1.0.0', '1.1.0', 'npm')
-
-      const analysis = await analyzeSecurityVulnerabilities(suspiciousDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      // Supply chain vulnerability might be detected by security patterns, not necessarily with specific type
-      const hasSecurityVuln = analysis.vulnerabilities.some(v =>
-        v.evidence?.some(
-          e => e.includes('malicious') || e.includes('backdoor') || e.includes('risk'),
-        ),
-      )
-      expect(hasSecurityVuln).toBe(true)
-    })
-
-    it('should detect typosquatting patterns', async () => {
-      const typoSquatDep = createMockDependency('reactt', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(typoSquatDep)
-
-      // Typosquatting detection might be basic, so we'll check if supply chain risks are detected
-      if (analysis.hasSecurityIssues) {
-        const supplyChainVuln = analysis.vulnerabilities.find(v => v.type === 'supply_chain')
-        expect(supplyChainVuln).toBeDefined()
-      }
-    })
-
-    it('should recognize trusted sources', async () => {
-      const trustedDep = createMockDependency('@types/node', '18.0.0', '18.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(trustedDep)
-
-      // Trusted packages should have lower risk scores but still may have 'low' severity
-      expect(analysis.overallSeverity).toMatch(/low|none/)
-      expect(analysis.recommendedAction).toBe('proceed')
-    })
-  })
-
-  describe('severity and risk calculation', () => {
-    it('should calculate appropriate risk scores', async () => {
-      const highRiskDep = createMockDependency('cve-2024-99999-critical', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(highRiskDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.overallSeverity).toMatch(/high|critical/)
-      expect(['low', 'medium', 'high']).toContain(analysis.confidence)
-    })
-
-    it('should provide appropriate confidence levels', async () => {
-      const unclearDep = createMockDependency('maybe-risky', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(unclearDep)
-
-      expect(typeof analysis.confidence).toBe('string') // Implementation returns string, not number
-      expect(['low', 'medium', 'high']).toContain(analysis.confidence)
-    })
-
-    it('should escalate severity appropriately', async () => {
-      const multipleCveDep = createMockDependency('cve-2024-1111-cve-2024-2222', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(multipleCveDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.cveCount).toBeGreaterThanOrEqual(2)
-      expect(analysis.overallSeverity).toMatch(/medium|high|critical/)
-    })
-  })
-
-  describe('recommendation engine', () => {
-    it('should recommend proceed for no security issues', async () => {
-      const safeDep = createMockDependency('@types/safe-package', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(safeDep)
-
-      expect(analysis.hasSecurityIssues).toBe(false)
-      expect(analysis.recommendedAction).toBe('proceed')
-      expect(analysis.reasoning).toContain('No security issues detected')
-    })
-
-    it('should recommend immediate action for critical vulnerabilities', async () => {
-      const criticalDep = createMockDependency('zero day exploit package', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(criticalDep)
-
-      expect(analysis.hasSecurityIssues).toBe(true)
-      expect(analysis.recommendedAction).toMatch(
-        /review_required|block_until_patched|immediate_update/,
-      )
-    })
-
-    it('should provide contextual reasoning', async () => {
-      const contextDep = createMockDependency('buffer overflow lib', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(contextDep)
-
-      expect(Array.isArray(analysis.reasoning)).toBe(true)
-      expect(analysis.reasoning.length).toBeGreaterThan(0)
-      expect(
-        analysis.reasoning.some(reason => typeof reason === 'string' && reason.length > 0),
-      ).toBe(true)
-    })
-  })
-
-  describe('edge cases and error handling', () => {
-    it('should handle empty dependency names', async () => {
-      const emptyDep = createMockDependency('', '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(emptyDep)
-
-      expect(analysis).toBeDefined()
-      expect(typeof analysis.hasSecurityIssues).toBe('boolean')
-    })
-
-    it('should handle unknown package managers', async () => {
-      const unknownDep = createMockDependency('some-package', '1.0.0', '1.1.0', 'unknown-manager')
-
-      const analysis = await analyzeSecurityVulnerabilities(unknownDep)
-
-      expect(analysis).toBeDefined()
-      expect(typeof analysis.hasSecurityIssues).toBe('boolean')
-    })
-
-    it('should handle malformed version strings', async () => {
-      const malformedDep = createMockDependency('test-package', 'not-a-version', 'also-not-version')
-
-      const analysis = await analyzeSecurityVulnerabilities(malformedDep)
-
-      expect(analysis).toBeDefined()
-      expect(typeof analysis.hasSecurityIssues).toBe('boolean')
-    })
-
-    it('should handle very long package names', async () => {
-      const longName = `${'a'.repeat(1000)}-malicious-package`
-      const longDep = createMockDependency(longName, '1.0.0', '1.1.0')
-
-      const analysis = await analyzeSecurityVulnerabilities(longDep)
-
-      expect(analysis).toBeDefined()
-      expect(typeof analysis.hasSecurityIssues).toBe('boolean')
-    })
-  })
-})
